Feat/layered water beds underwater vegetation

+++ src/models_3d/wikidata/placement.rs
@@ -992,7 +992,7 @@ mod tests {
         assert_eq!(material_to_rgb("nonsense"), None);
     }
 
-    fn block_ids(palette: &[Block]) -> Vec<u8> {
+    fn block_ids(palette: &[Block]) -> Vec<u16> {
         palette.iter().map(|b| b.id()).collect()
     }
